// ReactPortfolio/frontend/Webshop/src/core/models/Product.ts
export class Product {
    ID: number = 0;
    CategoryId: number = 0;
    CategoryName: string = "";
    Name: string = "";
    Price: number = 0;
    Currency: string = "";
    Count: number = 0;
    Description: string = "";
    Image: string = "";
    VisitCount: number = 0;
    ManufactureId: number=0;
    ManufacturerName:string="";
    Color:string="";

  }
  